+++ source/gui/nvdaControls.py
@@ -23,6 +23,7 @@ from config.featureFlag import (
 	FlagValueEnum as FeatureFlagEnumT,
 )
 import gui.message
+from winBindings import user32
 from .dpiScalingHelper import DpiScalingHelperMixin
 from . import (
 	guiHelper,
@@ -406,7 +407,7 @@ class EnhancedInputSlider(wx.Slider):
 		# HACK: Win events don't seem to be sent for certain explicitly set values,
 		# so send our own win event.
 		# This will cause duplicates in some cases, but NVDA will filter them out.
-		winUser.user32.NotifyWinEvent(
+		user32.NotifyWinEvent(
 			winUser.EVENT_OBJECT_VALUECHANGE,
 			self.Handle,
 			winUser.OBJID_CLIENT,